requirements.txt
langchain==0.3.13
langchain-community==0.3.12
langchain-openai==0.2.13
langchain-groq==0.2.2
langchain-core==0.3.27
langchain-nomic==0.1.4
pydantic==2.10.4
scikit-learn==1.6.0
pypdf==5.1.0
beautifulsoup4==4.12.3
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
faiss-cpu==1.9.0.post1
Jinja2==3.1.4
python-multipart==0.0.20